+++ versioned/transfer/src/main/java/org/projectnessie/versioned/transfer/NessieExporter.java
@@ -18,9 +18,12 @@ package org.projectnessie.versioned.transfer;
 import static org.projectnessie.versioned.transfer.ExportImportConstants.DEFAULT_BUFFER_SIZE;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.google.errorprone.annotations.CanIgnoreReturnValue;
 import jakarta.annotation.Nullable;
 import java.io.IOException;
+import java.net.URL;
 import java.time.Clock;
+import java.util.List;
 import org.immutables.value.Value;
 import org.projectnessie.versioned.StoreWorker;
 import org.projectnessie.versioned.VersionStore;
@@ -40,67 +43,88 @@ public abstract class NessieExporter {
 
   public static final String NAMED_REFS_PREFIX = "named-refs";
   public static final String COMMITS_PREFIX = "commits";
+  public static final String CUSTOM_PREFIX = "custom";
 
   public static Builder builder() {
     return ImmutableNessieExporter.builder();
   }
 
-  @SuppressWarnings("UnusedReturnValue")
   public interface Builder {
     /** Specify the {@code Persist} to use. */
+    @CanIgnoreReturnValue
     Builder persist(Persist persist);
 
+    @CanIgnoreReturnValue
     Builder commitLogic(CommitLogic commitLogic);
 
+    @CanIgnoreReturnValue
     Builder referenceLogic(ReferenceLogic referenceLogic);
 
+    @CanIgnoreReturnValue
     Builder repositoryLogic(RepositoryLogic repositoryLogic);
 
+    @CanIgnoreReturnValue
     Builder indexesLogic(IndexesLogic indexesLogic);
 
+    @CanIgnoreReturnValue
     Builder versionStore(VersionStore store);
 
     /** Optional, specify a custom {@link ObjectMapper}. */
+    @CanIgnoreReturnValue
     Builder objectMapper(ObjectMapper objectMapper);
 
     /** Optional, specify a custom {@link StoreWorker}. */
+    @CanIgnoreReturnValue
     Builder storeWorker(StoreWorker storeWorker);
 
     /**
      * Optional, specify a different buffer size than the default value of {@value
      * ExportImportConstants#DEFAULT_BUFFER_SIZE}.
      */
+    @CanIgnoreReturnValue
     Builder outputBufferSize(int outputBufferSize);
 
     /**
      * Maximum size of a file containing commits or named references. Default is to write everything
      * into a single file.
      */
+    @CanIgnoreReturnValue
     Builder maxFileSize(long maxFileSize);
 
     /**
      * The expected number of commits in the Nessie repository, default is {@value
      * ExportImportConstants#DEFAULT_EXPECTED_COMMIT_COUNT}.
      */
+    @CanIgnoreReturnValue
     Builder expectedCommitCount(int expectedCommitCount);
 
+    @CanIgnoreReturnValue
     Builder progressListener(ProgressListener progressListener);
 
+    @CanIgnoreReturnValue
     Builder exportFileSupplier(ExportFileSupplier exportFileSupplier);
 
+    @CanIgnoreReturnValue
     Builder fullScan(boolean fullScan);
 
+    @CanIgnoreReturnValue
     Builder contentsFromBranch(String branchName);
 
+    @CanIgnoreReturnValue
     Builder contentsBatchSize(int batchSize);
 
     /**
      * Optional, specify the number of commit log entries to be written at once, defaults to {@value
      * ExportImportConstants#DEFAULT_COMMIT_BATCH_SIZE}.
      */
+    @CanIgnoreReturnValue
     Builder commitBatchSize(int commitBatchSize);
 
+    @CanIgnoreReturnValue
     Builder exportVersion(int exportVersion);
+
+    @CanIgnoreReturnValue
+    Builder addGenericObjectResolvers(URL element);
 
     NessieExporter build();
   }
@@ -192,6 +216,8 @@ public abstract class NessieExporter {
     return ExportImportConstants.DEFAULT_EXPORT_VERSION;
   }
 
+  abstract List<URL> genericObjectResolvers();
+
   abstract ExportFileSupplier exportFileSupplier();
 
   @Value.Default
@@ -204,11 +230,12 @@ public abstract class NessieExporter {
 
     exportFiles.preValidate();
 
+    ExportVersion ver = ExportVersion.forNumber(exportVersion());
+
     if (contentsFromBranch() != null) {
-      return new ExportContents(exportFiles, this).exportRepo();
+      return new ExportContents(exportFiles, this, ver).exportRepo();
     }
 
-    return new ExportPersist(exportFiles, this, ExportVersion.forNumber(exportVersion()))
-        .exportRepo();
+    return new ExportPersist(exportFiles, this, ver).exportRepo();
   }
 }